use labels in double instead of float to make cpu training happy

+++ distilbert_data_utils.py
@@ -144,7 +144,7 @@ def convert_examples_to_features(
     input_id_tensor = torch.tensor(all_input_ids, dtype=torch.long)
     input_mask_tensor = torch.tensor(all_input_masks, dtype=torch.long)
     segment_id_tensor = torch.tensor(all_segment_ids, dtype=torch.long)
-    label_id_tensor = torch.tensor(all_label_ids, dtype=torch.float)
+    label_id_tensor = torch.tensor(all_label_ids, dtype=torch.double)
 
     return TensorDataset(
         input_id_tensor,
